requirements.txt
langchain
openai
python-dotenv
langchain-openai
ipykernel
beautifulsoup4
requests
langchain_community
faiss-cpu
streamlit
streamlit-chat
